Split tests out into specs and suites.
This change moves the tests out into  UMD modules containing BDD
specs. It introduces a test/unit.js suite to run the tests whilst
developing (browserified from the source) and a test/integration.js
to run the tests in CI, from the final build.

This also removes the need for the test/LegacySelleckt suite as that
was a poor replacement for proper integration tests.
