--- test/suite.js
@@ -1,2 +1,0 @@
-require('./SingleSelleckt.tests.js');
-require('./MultiSelleckt.tests.js');
